Extract `MarkdownWriter` to its own file
